New architecture for server Now requests must use content-length header to read from socket

+++ include/socket.h
@@ -11,6 +11,6 @@
 // Function headers:
 int close_socket(int);
 int connect_socket (int, struct sockaddr*, socklen_t);
-int read_socket(int, char*, size_t);
+ssize_t read_socket(int, char*, size_t);
 
 #endif // SOCKET_H
